refactor useBlockParentAttributes to ts
